Egress Rules Refactoring 2 (#1887)
* added model/egress_rules.go - put egress rules helper functions there

* fixed a typo

* use string formatting instead string concatenation

* use bytes.Buffer for effective string concatenation

* minor error string formatting fix

* remove a duplicated line
