+++ pilot/model/config.go
@@ -20,7 +20,6 @@ import (
 	"sort"
 
 	"github.com/golang/protobuf/proto"
-	multierror "github.com/hashicorp/go-multierror"
 
 	mccpb "istio.io/api/mixer/v1/config/client"
 	proxyconfig "istio.io/api/proxy/v1/config"
@@ -531,49 +530,6 @@ func (store *istioConfigStore) Policy(instances []*ServiceInstance, destination 
 	}
 
 	return &out
-}
-
-// RejectConflictingEgressRules rejects rules that have the destination which is equal to
-// the destionation of some other rule.
-// According to Envoy's virtual host specification, no virtual hosts can share the same domain.
-// The following code rejects conflicting rules deterministically, by a lexicographical order -
-// a rule with a smaller key lexicographically wins.
-// Here the key of the rule is the key of the Istio configuration objects - see
-// `func (meta *ConfigMeta) Key() string`
-func RejectConflictingEgressRules(egressRules map[string]*proxyconfig.EgressRule) ( // long line split
-	map[string]*proxyconfig.EgressRule, error) {
-	filteredEgressRules := make(map[string]*proxyconfig.EgressRule)
-	var errs error
-
-	var keys []string
-
-	// the key here is the key of the Istio configuration objects - see
-	// `func (meta *ConfigMeta) Key() string`
-	for key := range egressRules {
-		keys = append(keys, key)
-	}
-	sort.Strings(keys)
-
-	// domains - a map where keys are of the form domain:port and values are the keys of
-	// egress-rule configuration objects
-	domains := make(map[string]string)
-	for _, egressRuleKey := range keys {
-		egressRule := egressRules[egressRuleKey]
-		domain := egressRule.Destination.Service
-		keyOfAnEgressRuleWithTheSameDomain, conflictingRule := domains[domain]
-		if conflictingRule {
-			errs = multierror.Append(errs,
-				fmt.Errorf("rule %q conflicts with rule %q on domain "+
-					"%s, is rejected", egressRuleKey,
-					keyOfAnEgressRuleWithTheSameDomain, domain))
-			continue
-		}
-
-		domains[domain] = egressRuleKey
-		filteredEgressRules[egressRuleKey] = egressRule
-	}
-
-	return filteredEgressRules, errs
 }
 
 // `istio.mixer.v1.config.client.IstioService` and
